move TestSaveAssociationWithAutoIncrementField to association_test.go
