+++ tests/config_test.go
@@ -39,83 +39,20 @@
 package tests
 
 import (
-	"errors"
 	"regexp"
-	"sort"
 	"testing"
 
-	"github.com/google/uuid"
 	"github.com/oracle-samples/gorm-oracle/oracle"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
-	"gorm.io/gorm/utils/tests"
 )
 
-type FolderData struct {
-	ID         string           `gorm:"primaryKey;column:folder_id"`
-	Name       string           `gorm:"column:folder_nm"`
-	Properties []FolderProperty `gorm:"foreignKey:ID;PRELOAD:false"`
+type Student struct {
+	ID   uint
+	Name string
 }
 
-func (FolderData) TableName() string {
-	return "folder_data"
-}
-
-type FolderProperty struct {
-	Seq   uint64 `gorm:"autoIncrement"`
-	ID    string `gorm:"primaryKey;column:folder_id"`
-	Key   string `gorm:"primaryKey;unique"`
-	Value string
-}
-
-func (FolderProperty) TableName() string {
-	return "folder_property"
-}
-
-func TestSkipQuoteIdentifiersMigrator(t *testing.T) {
-	db, err := openTestDBWithOptions(
-		&oracle.Config{SkipQuoteIdentifiers: true},
-		&gorm.Config{Logger: newLogger})
-	if err != nil {
-		t.Fatalf("failed to connect database, got error %v", err)
-	}
-
-	db.Migrator().DropTable(&FolderData{}, &FolderProperty{})
-	db.Migrator().CreateTable(&FolderData{}, &FolderProperty{})
-
-	folderDataTN := "FOLDER_DATA"
-	if !db.Migrator().HasTable(folderDataTN) {
-		t.Errorf("Failed to get table: %s", folderDataTN)
-	}
-
-	if !db.Migrator().HasColumn(folderDataTN, "FOLDER_ID") {
-		t.Errorf("Failed to get column: FOLDER_ID")
-	}
-
-	if !db.Migrator().HasColumn(folderDataTN, "FOLDER_NM") {
-		t.Errorf("Failed to get column: FOLDER_NM")
-	}
-
-	folderPropertyTN := "FOLDER_PROPERTY"
-	if !db.Migrator().HasTable(folderPropertyTN) {
-		t.Errorf("Failed to get table: %s", folderPropertyTN)
-	}
-
-	if !db.Migrator().HasColumn(folderPropertyTN, "SEQ") {
-		t.Errorf("Failed to get column: SEQ")
-	}
-
-	if !db.Migrator().HasColumn(folderPropertyTN, "FOLDER_ID") {
-		t.Errorf("Failed to get column: FOLDER_ID")
-	}
-
-	if !db.Migrator().HasColumn(folderPropertyTN, "KEY") {
-		t.Errorf("Failed to get column: KEY")
-	}
-
-	if !db.Migrator().HasColumn(folderPropertyTN, "VALUE") {
-		t.Errorf("Failed to get column: VALUE")
-	}
+func (s Student) TableName() string {
+	return "STUDENTS"
 }
 
 func TestSkipQuoteIdentifiers(t *testing.T) {
@@ -126,95 +63,41 @@ func TestSkipQuoteIdentifiers(t *testing.T) {
 		t.Fatalf("failed to connect database, got error %v", err)
 	}
 
-	db.Migrator().DropTable(&FolderData{}, &FolderProperty{})
-	db.Migrator().CreateTable(&FolderData{}, &FolderProperty{})
+	db.Migrator().DropTable(&Student{})
+	db.Migrator().CreateTable(&Student{})
 
-	id := uuid.New().String()
-	folder := FolderData{
-		ID:   id,
-		Name: "My Folder",
-		Properties: []FolderProperty{
-			{
-				ID:    id,
-				Key:   "foo1",
-				Value: "bar1",
-			},
-			{
-				ID:    id,
-				Key:   "foo2",
-				Value: "bar2",
-			},
-		},
+	if !db.Migrator().HasTable(&Student{}) {
+		t.Errorf("Failed to get table: student")
 	}
 
-	if err := db.Create(&folder).Error; err != nil {
-		t.Errorf("Failed to insert data, got %v", err)
+	if !db.Migrator().HasColumn(&Student{}, "ID") {
+		t.Errorf("Failed to get column: id")
 	}
 
-	createdFolder := FolderData{}
-	if err := db.Model(&FolderData{}).Preload("Properties").First(&createdFolder).Error; err != nil {
-		t.Errorf("Failed to query data, got %v", err)
+	if !db.Migrator().HasColumn(&Student{}, "NAME") {
+		t.Errorf("Failed to get column: name")
 	}
 
-	CheckFolderData(t, createdFolder, folder)
-
-	createdFolder.Properties[1].Value = "baz1"
-	createdFolder.Properties = append(createdFolder.Properties, FolderProperty{
-		ID:    id,
-		Key:   "foo3",
-		Value: "bar3",
-	})
-	createdFolder.Properties = append(createdFolder.Properties, FolderProperty{
-		ID:    id,
-		Key:   "foo4",
-		Value: "bar4",
-	})
-	db.Save(&createdFolder)
-
-	updatedFolder := FolderData{}
-	if err := db.Model(&FolderData{}).Preload("Properties").First(&updatedFolder).Error; err != nil {
-		t.Errorf("Failed to query data, got %v", err)
+	student := Student{ID: 1, Name: "John"}
+	if err := db.Model(&Student{}).Create(&student).Error; err != nil {
+		t.Errorf("Failed to insert student, got %v", err)
 	}
 
-	CheckFolderData(t, updatedFolder, createdFolder)
-
-	if err := db.Select(clause.Associations).Delete(&createdFolder).Error; err != nil {
-		t.Errorf("Failed to delete data, got %v", err)
+	var result Student
+	if err := db.First(&result).Error; err != nil {
+		t.Errorf("Failed to query first student, got %v", err)
 	}
 
-	result := FolderData{}
-	if err := db.Where("folder_id = ?", createdFolder.ID).First(&result).Error; err == nil || !errors.Is(err, gorm.ErrRecordNotFound) {
-		t.Errorf("should returns record not found error, but got %v", err)
+	if result.ID != student.ID {
+		t.Errorf("id should be %v, but got %v", student.ID, result.ID)
 	}
-}
 
-func CheckFolderData(t *testing.T, folderData FolderData, expect FolderData) {
-	tests.AssertObjEqual(t, folderData, expect, "ID", "Name")
-	t.Run("Properties", func(t *testing.T) {
-		if len(folderData.Properties) != len(expect.Properties) {
-			t.Fatalf("properties should equal, expect: %v, got %v", len(expect.Properties), len(folderData.Properties))
-		}
-
-		sort.Slice(folderData.Properties, func(i, j int) bool {
-			return folderData.Properties[i].ID > folderData.Properties[j].ID
-		})
-
-		sort.Slice(expect.Properties, func(i, j int) bool {
-			return expect.Properties[i].ID > expect.Properties[j].ID
-		})
-
-		for idx, property := range folderData.Properties {
-			tests.AssertObjEqual(t, property, expect.Properties[idx], "Seq", "ID", "Key", "Value")
-		}
-	})
+	if result.Name != student.Name {
+		t.Errorf("name should be %v, but got %v", student.Name, result.Name)
+	}
 }
 
 func TestSkipQuoteIdentifiersSQL(t *testing.T) {
-	type Student struct {
-		ID   uint
-		Name string
-	}
-
 	db, err := openTestDBWithOptions(
 		&oracle.Config{SkipQuoteIdentifiers: true},
 		&gorm.Config{Logger: newLogger})
@@ -226,7 +109,7 @@ func TestSkipQuoteIdentifiersSQL(t *testing.T) {
 	insertedStudent := Student{ID: 1, Name: "John"}
 	result := dryrunDB.Model(&Student{}).Create(&insertedStudent)
 
-	if !regexp.MustCompile(`^INSERT INTO students \(name,id\) VALUES \(:1,:2\)$`).MatchString(result.Statement.SQL.String()) {
+	if !regexp.MustCompile(`^INSERT INTO STUDENTS \(name,id\) VALUES \(:1,:2\)$`).MatchString(result.Statement.SQL.String()) {
 		t.Errorf("invalid insert SQL, got %v", result.Statement.SQL.String())
 	}
 
@@ -234,26 +117,26 @@ func TestSkipQuoteIdentifiersSQL(t *testing.T) {
 	var firstStudent Student
 	result = dryrunDB.First(&firstStudent)
 
-	if !regexp.MustCompile(`^SELECT \* FROM students ORDER BY students\.id FETCH NEXT 1 ROW ONLY$`).MatchString(result.Statement.SQL.String()) {
+	if !regexp.MustCompile(`^SELECT \* FROM STUDENTS ORDER BY STUDENTS\.id FETCH NEXT 1 ROW ONLY$`).MatchString(result.Statement.SQL.String()) {
 		t.Fatalf("SQL should include selected names, but got %v", result.Statement.SQL.String())
 	}
 
 	// Test Find
 	var foundStudent Student
 	result = dryrunDB.Find(foundStudent, "id = ?", insertedStudent.ID)
-	if !regexp.MustCompile(`^SELECT \* FROM students WHERE id = :1$`).MatchString(result.Statement.SQL.String()) {
+	if !regexp.MustCompile(`^SELECT \* FROM STUDENTS WHERE id = :1$`).MatchString(result.Statement.SQL.String()) {
 		t.Fatalf("SQL should include selected names, but got %v", result.Statement.SQL.String())
 	}
 
 	// Test Save
 	result = dryrunDB.Save(&Student{ID: 2, Name: "Mary"})
-	if !regexp.MustCompile(`^UPDATE students SET name=:1 WHERE id = :2$`).MatchString(result.Statement.SQL.String()) {
+	if !regexp.MustCompile(`^UPDATE STUDENTS SET name=:1 WHERE id = :2$`).MatchString(result.Statement.SQL.String()) {
 		t.Fatalf("SQL should include selected names, but got %v", result.Statement.SQL.String())
 	}
 
 	// Update with conditions
 	result = dryrunDB.Model(&Student{}).Where("id = ?", 1).Update("name", "hello")
-	if !regexp.MustCompile(`^UPDATE students SET name=:1 WHERE id = :2$`).MatchString(result.Statement.SQL.String()) {
+	if !regexp.MustCompile(`^UPDATE STUDENTS SET name=:1 WHERE id = :2$`).MatchString(result.Statement.SQL.String()) {
 		t.Fatalf("SQL should include selected names, but got %v", result.Statement.SQL.String())
 	}
 }
